+++ test/cljc/matthiasn/systems_toolbox/test.cljs
@@ -1,5 +1,0 @@
-(ns matthiasn.systems-toolbox.test)
-
-(enable-console-print!)
-
-(prn :foo)