+++ test/cljs/matthiasn/systems_toolbox/runner.cljs
@@ -1,7 +1,0 @@
-(ns matthiasn.systems-toolbox.runner
-  (:require [doo.runner :refer-macros [doo-tests]]
-            [matthiasn.systems-toolbox.test]
-            [matthiasn.systems-toolbox.component-test]))
-
-(doo-tests 'matthiasn.systems-toolbox.test
-           'matthiasn.systems-toolbox.component-test)