+++ test/cljs/matthiasn/systems_toolbox/test.cljs
@@ -1,5 +1,0 @@
-(ns matthiasn.systems-toolbox.test
-  (:require [cljs.test :refer-macros [deftest is]]))
-
-(deftest do-i-work
-  (is (= 2 2)))