+++ test/matthiasn/systems_toolbox/log.cljs
@@ -1,0 +1,5 @@
+(ns matthiasn.systems-toolbox.log)
+
+(defn debug
+  [& args]
+  (prn args))